Add tuner for confidence threshold.

+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/behaviorTree/action/actions/SceneActionDefinition.java
@@ -296,7 +296,7 @@ public class SceneActionDefinition extends ActionNodeDefinition
       return nominalObjectPose;
    }
 
-   public double getYoloConfidenceThreshold()
+   public float getYoloConfidenceThreshold()
    {
       return yoloConfidenceThreshold.getValue();
    }
